+++ src/solvers/gradient_descent.jl
@@ -80,7 +80,7 @@ function (r::IdentityUpdateRule)(
         mp::AbstractManoptProblem, s::AbstractGradientSolverState, k
     )
     get_gradient!(mp, s.X, s.p)
-    return get_stepsize(mp, s, k; gradient = s.X), s.X
+    return get_stepsize(mp, s, k, -s.X; gradient = s.X), s.X
 end
 
 function default_stepsize(